Replace custom `IrFactory` with the `DeclarationIrBuilder`
Tools to construct IRs are spread across multiple types:

- `IrElementBuilder` are low-level builders that can construct isolated IRs (e.g., functions). The consumer needs to specify every single argument to create the IR (e.g., function name, its position, its return type, etc). Such builders are unaware of the scope they are in and the type system.
- `IrFactory` provides abstractions over `IrElementBuilder`s to simplify IR creation. Similarly to the individual builders, the factory is unaware of the scope (e.g., the class it is in when creating a function) and the type system
- `IrBuilder`s is scope- and type-aware `IrFactory` with a wider range of utility methods that simplify the IR creation even further. The `IrBuilder` is type- (i.e., can reference the `IrBuiltIns`) and is scope-aware (i.e., it can infer some of the required `IrElement` attributes from the current scope). So, effectively it provides the same functionality as the `ContextAwareIrFactory` with additional utility functions.

For example, it is handy when we need to construct dependent nested IRs (e.g., body of a function). To create a body, we need to attach it to a function symbol. In case of a factory, we need to pass the symbol around and also ensure that some constraints are met (e.g., return types and code positions). With the `IrBuilder`, we can create a single builder instance that is aware of the function scope and let it propagate this info to all builder functions.